Refactor utility functions
Move several utility functions to different files to tidy the main
revtui file.
